refactor(filid): group cacheManager cache helpers into caches/utils organ
Move domain-neutral helpers out of caches/ root into a new caches/utils/
sub-organ: the flat util files (cwdHash, sessionIdHash, getCacheDir,
getPluginRoot, fmapPath+VisitScope) plus per-file inline helpers extracted
from commitVisit (writeAtomic, canonicalOf), fmapLock (ownerPath), and
turnCounter (turnPath). Consumers switch to ./utils/* imports; facade
re-exports and public signatures are unchanged, and hooks reach these
through the facade so bundles stay tree-shaken. INTENT.md structure refreshed.
