+++ plugins/filid/src/core/infra/cacheManager/caches/fractalMapCache.ts
@@ -1,8 +1,8 @@
 export type { FractalMap } from './readFractalMap.js';
 export { readFractalMap } from './readFractalMap.js';
 export { removeFractalMap } from './removeFractalMap.js';
-export type { VisitScope } from './fmapPath.js';
-export { fmapPath } from './fmapPath.js';
+export type { VisitScope } from './utils/fcaMapPath.js';
+export { fcaMapPath } from './utils/fcaMapPath.js';
 export { commitVisit } from './commitVisit.js';
 export type {
   DeliveredState,
