+++ plugins/filid/src/core/infra/cacheManager/caches/isFirstInSession.ts
@@ -3,8 +3,8 @@ import { join } from 'node:path';
 
 import { createLogger } from '../../../../lib/logger.js';
 
-import { getCacheDir } from './getCacheDir.js';
-import { sessionIdHash } from './sessionIdHash.js';
+import { getCacheDir } from './utils/getCacheDir.js';
+import { sessionIdHash } from './utils/sessionIdHash.js';
 
 const log = createLogger('cache');
