+++ plugins/filid/src/core/infra/cacheManager/caches/isSessionPruneDue.ts
@@ -7,7 +7,7 @@ import {
 } from '../../../../constants/infraDefaults.js';
 import { createLogger } from '../../../../lib/logger.js';
 
-import { getCacheDir } from './getCacheDir.js';
+import { getCacheDir } from './utils/getCacheDir.js';
 
 const log = createLogger('cache');
