+++ plugins/filid/src/core/infra/cacheManager/caches/markPruneRun.ts
@@ -4,7 +4,7 @@ import { join } from 'node:path';
 import { PRUNE_MARKER_FILENAME } from '../../../../constants/infraDefaults.js';
 import { createLogger } from '../../../../lib/logger.js';
 
-import { getPluginRoot } from './getPluginRoot.js';
+import { getPluginRoot } from './utils/getPluginRoot.js';
 
 const log = createLogger('cache');
