+++ plugins/filid/src/core/infra/cacheManager/caches/sessionCache.ts
@@ -5,9 +5,9 @@
  * Kept so existing importers (cacheManager.ts facade, sibling caches/*.ts)
  * continue to reference `./sessionCache.js` unchanged.
  */
-export { cwdHash } from './cwdHash.js';
-export { getCacheDir } from './getCacheDir.js';
-export { getPluginRoot } from './getPluginRoot.js';
+export { cwdHash } from './utils/cwdHash.js';
+export { getCacheDir } from './utils/getCacheDir.js';
+export { getPluginRoot } from './utils/getPluginRoot.js';
 export { isFirstInSession } from './isFirstInSession.js';
 export { isPruneDue } from './isPruneDue.js';
 export { isSessionPruneDue } from './isSessionPruneDue.js';
@@ -17,4 +17,4 @@ export { markSessionPruneRun } from './markSessionPruneRun.js';
 export { pruneOldSessions } from './pruneOldSessions.js';
 export { pruneStaleCacheDirs } from './pruneStaleCacheDirs.js';
 export { removeSessionFiles } from './removeSessionFiles.js';
-export { sessionIdHash } from './sessionIdHash.js';
+export { sessionIdHash } from './utils/sessionIdHash.js';
