+++ plugins/filid/src/core/infra/cacheManager/caches/utils/cwdHash.ts
