+++ plugins/filid/src/core/infra/cacheManager/caches/utils/getPluginRoot.ts
