+++ plugins/filid/src/core/infra/cacheManager/caches/utils/ownerPath.ts
@@ -1,0 +1,6 @@
+import { join } from 'node:path';
+
+/** Ownership-token file inside the lock dir; distinguishes lock generations. */
+export function ownerPath(lockPath: string): string {
+  return join(lockPath, 'owner');
+}
