+++ plugins/filid/src/core/infra/cacheManager/caches/utils/sessionIdHash.ts
