+++ plugins/filid/src/core/infra/cacheManager/caches/utils/writeAtomic.ts
@@ -1,0 +1,8 @@
+import { renameSync, writeFileSync } from 'node:fs';
+
+/** Crash-safe write: pid-scoped tmp file swapped into place via atomic rename. */
+export function writeAtomic(filePath: string, data: string): void {
+  const tmpPath = `${filePath}.${process.pid}.tmp`;
+  writeFileSync(tmpPath, data);
+  renameSync(tmpPath, filePath);
+}
